# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Universal Document Vector Search Service",
    "description": "Микросервис для векторного поиска по документам с поддержкой метаданных, категорий и версионирования",
    "version": "1.2.0"
  },
  "paths": {
...
    "/v1/admin/hnsw": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Включить/выключить HNSW-индекс",
        "description": "Включает или выключает HNSW-индексацию для указанной коллекции.\n\nКак использовать:\n- **Выключить** HNSW (ускоряет массовую загрузку): `enabled=false` (эквивалент `m=0`)\n- **Включить** HNSW (ускоряет поиск): `enabled=true` и задать `m` (обычно 16)\n\nПараметр `m` определяет количество связей между узлами (чем больше, тем выше качество/память).\nHNSW (Hierarchical Navigable Small World) — индекс для ускорения поиска по векторам в Qdrant.",
        "operationId": "set_hnsw_v1_admin_hnsw_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_set_hnsw_v1_admin_hnsw_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/admin/create_payload_index": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Создать индекс для поля payload",
        "description": "Создаёт индекс для указанного поля в payload коллекции.\n\n**Поддерживаемые типы:** `keyword`, `integer`, `float`, `geo`.\n\nИндексация ускоряет фильтрацию при поиске (например, по `source_id`, `category_path`).\nРекомендуется индексировать поля, по которым часто выполняется фильтрация.",
        "operationId": "create_payload_index_v1_admin_create_payload_index_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayloadIndexRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/admin/collections": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get Collections With Categories",
        "operationId": "get_collections_with_categories_v1_admin_collections_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionsWithCategoriesResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/admin/categories/hierarchy": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Иерархия категорий с количеством чанков",
        "description": "Возвращает категории с количеством чанков для указанной коллекции.\n\n**Параметры:**\n- `collection_name`: имя коллекции (опционально, если не указан - возвращает категории для всех коллекций)\n- `depth`: количество уровней для получения (0, 1-10)\n  - `0` - только список коллекций без категорий\n  - `1-10` - иерархия категорий (максимум 10 уровней)\n- `categories`: массив полных путей категорий (опционально)\n  - Если указаны - возвращает только коллекции, содержащие эти категории\n  - Если не указаны - возвращает все категории для указанных/всех коллекций\n\n**Логика:**\n- Если `categories` не указаны и `depth=0` - возвращает список коллекций без категорий\n- Если `categories` не указаны и `depth>0` - возвращает первые `depth` уровней\n- Если `categories` указаны - для каждого пути возвращает `depth` вложенных уровней\n- Если указана `collection_name` и она не существует - возвращает ошибку 404\n\n**Пример ответа (depth=1, categories не указаны):**\n```json\n{\n  \"results\": [\n    {\"name\": \"collection1\", \"categories\": [\n      {\"category_path\": \"Docs\", \"chunk_count\": 150},\n      {\"category_path\": \"API\", \"chunk_count\": 50}\n    ]}\n  ]\n}\n```\n\n**Пример ответа (depth=0):**\n```json\n{\n  \"results\": [\n    {\"name\": \"collection1\", \"categories\": []},\n    {\"name\": \"collection2\", \"categories\": []}\n  ]\n}\n```\n\n**Пример ответа (categories указанны, несуществующая категория):**\n```json\n{\n  \"results\": []\n}\n```\n\n**Ошибки:**\n- 404 - коллекция не найдена (если указана `collection_name`)\n- 422 - неверные параметры (например, `depth > 10`)",
        "operationId": "get_category_hierarchy_v1_admin_categories_hierarchy_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryHierarchyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryHierarchyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_set_hnsw_v1_admin_hnsw_post": {
        "properties": {
          "collection_name": {
            "type": "string",
            "title": "Collection Name",
            "description": "Имя коллекции"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "description": "Включить (true) или выключить (false) HNSW"
          },
          "m": {
            "type": "integer",
            "minimum": 0.0,
            "title": "M",
            "description": "Параметр HNSW m (используется только при enabled=true)",
            "default": 16
          }
        },
        "type": "object",
        "required": [
          "collection_name",
          "enabled"
        ],
        "title": "Body_set_hnsw_v1_admin_hnsw_post"
      },
...
      "CategoryChunkCount": {
        "properties": {
          "category_path": {
            "type": "string",
            "title": "Category Path"
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count"
          }
        },
        "type": "object",
        "required": [
          "category_path",
          "chunk_count"
        ],
        "title": "CategoryChunkCount"
      },
      "CategoryHierarchyRequest": {
        "properties": {
          "collection_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collection Name",
            "description": "Имя коллекции (опционально)"
          },
          "depth": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Depth",
            "description": "Количество уровней для получения (0 - только список коллекций без категорий, 1-10 - иерархия категорий). Максимальное значение: 10.",
            "default": 0
          },
          "categories": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Categories",
            "description": "Массив полных путей категорий (опционально). Если указаны, возвращаются только коллекции с этими категориями."
          }
        },
        "type": "object",
        "title": "CategoryHierarchyRequest"
      },
      "CategoryHierarchyResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/CollectionCategoryHierarchy"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "results"
        ],
        "title": "CategoryHierarchyResponse"
      },
...
      "CollectionCategories": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "categories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Categories"
          }
        },
        "type": "object",
        "required": [
          "name",
          "categories"
        ],
        "title": "CollectionCategories"
      },
      "CollectionCategoryHierarchy": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/CategoryChunkCount"
            },
            "type": "array",
            "title": "Categories"
          }
        },
        "type": "object",
        "required": [
          "name",
          "categories"
        ],
        "title": "CollectionCategoryHierarchy"
      },
      "CollectionsWithCategoriesResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/CollectionCategories"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "results"
        ],
        "title": "CollectionsWithCategoriesResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PayloadIndexRequest": {
        "properties": {
          "collection_name": {
            "type": "string",
            "title": "Collection Name"
          },
          "field_name": {
            "type": "string",
            "title": "Field Name"
          },
          "field_type": {
            "$ref": "#/components/schemas/PayloadSchemaType",
            "default": "keyword"
          }
        },
        "type": "object",
        "required": [
          "collection_name",
          "field_name"
        ],
        "title": "PayloadIndexRequest"
      },
      "PayloadSchemaType": {
        "type": "string",
        "enum": [
          "keyword",
          "integer",
          "float",
          "geo",
          "text",
          "bool",
          "datetime",
          "uuid"
        ],
        "title": "PayloadSchemaType",
        "description": "All possible names of payload types"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      },
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}